# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Calculator Service
    description: Service for multiplying numbers, a Goa teaser
    version: 0.0.1
host: localhost:8000
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /div/{c}/{d}:
        get:
            tags:
                - calc
            summary: divide calc
            description: Divide returns the integral division of two integers.
            operationId: calc#divide
            parameters:
                - name: c
                  in: path
                  description: Left operand
                  required: true
                  type: integer
                - name: d
                  in: path
                  description: Right operand
                  required: true
                  type: integer
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: integer
                        format: int64
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/CalcDivideDivByZeroResponseBody'
            schemes:
                - http
    /multiply/{a}/{b}:
        get:
            tags:
                - calc
            summary: multiply calc
            operationId: calc#multiply
            parameters:
                - name: a
                  in: path
                  description: Left operand
                  required: true
                  type: integer
                - name: b
                  in: path
                  description: Right operand
                  required: true
                  type: integer
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: integer
                        format: int64
            schemes:
                - http
    /update:
        put:
            tags:
                - calc
            summary: update calc
            description: Change account name
            operationId: calc#update
            parameters:
                - name: UpdateRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/CalcUpdateRequestBody'
                    required:
                        - name
                        - email
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/CalcUpdateResponseBody'
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/CalcUpdateNotFoundResponseBody'
            schemes:
                - http
definitions:
    CalcDivideDivByZeroResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: true
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: true
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: true
        description: divide_DivByZero_response_body result type (default view)
        example:
            fault: true
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: true
            timeout: true
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
    CalcUpdateNotFoundResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: false
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: true
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: true
        description: update_NotFound_response_body result type (default view)
        example:
            fault: false
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: true
            timeout: true
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
    CalcUpdateRequestBody:
        title: CalcUpdateRequestBody
        type: object
        properties:
            email:
                type: string
                example: kaleb.quitzon@ortizboyer.info
                format: email
            name:
                type: string
                example: m5h
                minLength: 3
                maxLength: 20
        example:
            email: alejandrin@bradtke.info
            name: u7e
        required:
            - name
            - email
    CalcUpdateResponseBody:
        title: 'Mediatype identifier: application/vnd.create; view=default'
        type: object
        properties:
            href:
                type: string
                description: Href of the created resource
                example: Dolore qui consequuntur qui dolor.
            name:
                type: string
                description: Name of the created resource
                example: Omnis dolores accusantium doloribus qui.
        description: UpdateResponseBody result type (default view)
        example:
            href: Dolores voluptas suscipit ut rerum odio.
            name: Ipsum nostrum cumque deserunt labore hic rerum.